LRDOCS-11747 renames workspace to liferay-courses-workspace
