--- workspaces/liferay-course-workspace/client-extensions/liferay-course-etc-spring-boot/client-extension.yaml
@@ -1,22 +1,0 @@
-assemble:
-    - fromTask: bootJar
-liferay-course-etc-spring-boot-oauth-application-user-agent:
-    .serviceAddress: localhost:58081
-    .serviceScheme: http
-    name: Liferay Course Etc Spring Boot Spring Boot OAuth Application User Agent
-    scopes:
-        - Liferay.Headless.Admin.User.everything 
-        - Liferay.Headless.Admin.Workflow.everything
-    type: oAuthApplicationUserAgent
-liferay-course-etc-spring-boot-object-action-1:
-    name: Liferay Course Etc Spring Boot Spring Boot Object Action 1
-    oAuth2ApplicationExternalReferenceCode: liferay-course-etc-spring-boot-oauth-application-user-agent
-    resourcePath: /object/action/1
-    type: objectAction
-    allowedObjectDefinitionNames:
-        - C_DistributorApplication
-liferay-course-etc-spring-boot-workflow-action-1:
-    name: Liferay Course Etc Spring Boot Spring Boot Workflow Action 1
-    oAuth2ApplicationExternalReferenceCode: liferay-course-etc-spring-boot-oauth-application-user-agent
-    resourcePath: /workflow/action/1
-    type: workflowAction
