--- workspaces/liferay-course-workspace/poshi/poshi.properties
@@ -1,2 +1,0 @@
-test.base.dir.name=.
-#test.name=LiferayCourseSmoke
