+++ workspaces/liferay-courses-workspace/README.markdown
@@ -1,0 +1,26 @@
+# Liferay Courses Workspace
+
+The Liferay Courses Workspace contains sample [client extensions](https://learn.liferay.com/web/guest/w/dxp/building-applications/client-extensions) for use with [Liferay Learn courses](https://learn.liferay.com/web/guest/w/courses/index). To use them, follow along with the corresponding course:
+
+- [Application Development](#application-development)
+- [Site Designer](#site-designer)
+
+## Application Development
+
+The [Application Development](https://learn.liferay.com/web/guest/w/courses/application-development) course covers the development paradigm for extending Liferay in 7.4+. While these lessons are applicable to Liferay Experience Cloud (LXC), LXC Self-Managed, and Liferay DXP Self-Hosted.
+
+This course consists of three modules that guide users through building a distributor onboarding flow using Liferay's low-code features and client extensions.
+
+### List of Client Extension Projects
+
+- *liferay-courses-batch*
+
+   Includes a workflow definition for use with the `liferay-courses-etc-spring-boot` workflow action.
+
+- *liferay-courses-etc-spring-boot*
+
+   Includes two microservice client extensions: workflow action and object action. You can only use them with the Distributor Application object. Also, the workflow action depends on the `liferay-courses-batch` workflow definition.
+
+## Site Designer
+
+Coming Soon!
