+++ workspaces/liferay-courses-workspace/client-extensions/liferay-courses-etc-spring-boot/Dockerfile
