+++ workspaces/liferay-courses-workspace/client-extensions/liferay-courses-etc-spring-boot/LCP.json
@@ -5,7 +5,7 @@
       "deploy": false
     }
   },
-  "id": "liferaycourseetcspringboot",
+  "id": "liferaycoursesetcspringboot",
   "kind": "Deployment",
   "livenessProbe": {
     "httpGet": {
