+++ workspaces/liferay-courses-workspace/client-extensions/liferay-courses-etc-spring-boot/src/main/java/com/liferay/courses/ObjectAction1RestController.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.course;
+package com.liferay.courses;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
