+++ workspaces/liferay-courses-workspace/client-extensions/liferay-courses-etc-spring-boot/src/main/resources/application-default.properties
@@ -10,7 +10,7 @@ com.liferay.lxc.dxp.server.protocol=http
 # OAuth
 #
 
-liferay.oauth.application.external.reference.codes=liferay-course-etc-spring-boot-oauth-application-user-agent
+liferay.oauth.application.external.reference.codes=liferay-courses-etc-spring-boot-oauth-application-user-agent
 liferay.oauth.urls.excludes=/ready
 
 #
