+++ workspaces/liferay-courses-workspace/client-extensions/liferay-courses-etc-spring-boot/src/main/resources/application.properties
