+++ workspaces/liferay-courses-workspace/configs/local/portal-ext.properties
