+++ workspaces/liferay-courses-workspace/gradle.properties
