+++ workspaces/liferay-courses-workspace/poshi/poshi.properties
@@ -1,0 +1,2 @@
+test.base.dir.name=.
+#test.name=LiferayCoursesSmoke
