+++ workspaces/liferay-courses-workspace/settings.gradle
