+++ workspaces/liferay-courses-workspace/test.properties
@@ -7,13 +7,13 @@
     #
 
     test.batch.run.property.query[functional-tomcat*-mysql*-jdk8][relevant]=\
-        (testray.main.component.name ~ "Liferay Course Workspace")
+        (testray.main.component.name ~ "Liferay Courses Workspace")
 
     test.batch.run.property.query[functional-workspaces-bundle-mysql*-jdk8][relevant]=\
-        (testray.main.component.name ~ "Liferay Course Workspace")
+        (testray.main.component.name ~ "Liferay Courses Workspace")
 
 ##
 ## Testray
 ##
 
-    testray.main.component.name=Liferay Course Workspace
+    testray.main.component.name=Liferay Courses Workspace
